Bildirim/backendbilidirm/requirements.txt
# requirements.txt
firebase-admin==6.2.0
sendgrid==6.10.0
twilio==8.10.0
python-dotenv==1.0.0
psycopg2-binary==2.9.7
sqlalchemy==2.0.21
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
redis==5.0.1
celery==5.3.4
pytest==7.4.3
requests==2.31.0
python-multipart==0.0.6